requirements.txt
colorama==0.4.1
coloredlogs==14.0
numpy==1.18.5
Pillow==7.1.2
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.8.0
typer==0.3.1